+++ base_extended_som/__init__.py
@@ -1,0 +1,1 @@
+import res_partner
