+++ www_som/__terp__.py
@@ -9,9 +9,12 @@ Mòdul per la integració de l'oficina virtual
     "category": "www",
     "depends": [
         "base_extended",
+        "base_extended_som",
         "www_base",
         "som_polissa_soci",
         "som_indexada",
+        "som_infoenergia",
+        "poweremail",
         "giscedata_lectures_comer",
         "giscedata_lectures_pool",
         "giscedata_facturacio_impagat_comer",
@@ -21,6 +24,7 @@ Mòdul per la integració de l'oficina virtual
     "demo_xml": [
         "demo/res_partner_demo.xml",
         "demo/indexed_prices_demo.xml",
+        "demo/ov_attachment_demo.xml",
     ],
     "update_xml": [
         "www_som_data.xml",
