+++ www_som/tests/__init__.py
@@ -1,3 +1,4 @@
 from tests_polissa import *
 from tests_factura import *
 from tests_indexada_helpers import *
+from tests_partner import *
